Combine all code into a single module
